requirements.txt
faiss-cpu==1.7.4
sentence-transformers==2.2.2
llama-cpp-python==0.2.61
PyPDF2==3.0.1
numpy==1.26.4
huggingface_hub==0.14.1
flask==2.3.2
waitress